[feat] 마커에 로고, 홈페이지 주소 추가
- json파일에 image, home_url 스크래핑 후 데이터 추가
- 화면에 json data 랜더링(imgae, home_url)
